Use upstream Intuition with Composer
This commit removes the customized copy of Intuition from the repo, and
adds the upstream Intuition framework as a Composer dependency. Looking at
the recent changes of the Intuition framework, I believe it's now usable
for reFill.

This is part of the efforts to move the translation workflow back to
Translatewiki.